:recycle: Update Auth style
