+++ .prettierrc.json
@@ -1,8 +1,8 @@
 {
-    "tabWidth": 2,
-    "useTabs": false,
-    "semi": false,
-    "singleQuote": true,
-    "jsxSingleQuote": true,
-    "arrowParens": "always"
-  }
+  "tabWidth": 2,
+  "useTabs": false,
+  "semi": false,
+  "singleQuote": true,
+  "jsxSingleQuote": true,
+  "arrowParens": "always"
+}
