+++ shared/common/src/pages/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginPage } from './login'
+export { default as SignUpPage } from './signUp'
