+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpPage1/index.tsx
