+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpPage2/index.tsx
