+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpPage3/index.tsx
