+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpScrollContainer/style.ts
