+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpSuccess/index.stories.tsx
