+++ shared/common/src/pages/auth/signUp/SignUpPages/SignUpSuccess/style.ts
