+++ shared/common/src/pages/auth/signUp/SignUpPages/index.ts
