+++ shared/common/src/pages/auth/signUp/SignUpTitle/style.ts
