+++ shared/common/src/pages/auth/signUp/index.tsx
@@ -2,15 +2,15 @@
 
 import { SignUpPageNumber } from '@bitgouel/common'
 import { useRecoilValue } from 'recoil'
-import { match } from 'ts-pattern'
+import { AuthWrapper } from '../style'
 import {
   SignUpPage1,
   SignUpPage2,
   SignUpPage3,
   SignUpSuccess,
 } from './SignUpPages'
-import * as S from './style'
 import SignUpTitle from './SignUpTitle'
+import * as S from './style'
 
 const SignUpPage = () => {
   const signUpPage = useRecoilValue(SignUpPageNumber)
@@ -19,14 +19,14 @@ const SignUpPage = () => {
     <>
       {signUpPage === 4 && <SignUpSuccess />}
       {signUpPage < 4 && (
-        <S.SignUpWrapper>
+        <AuthWrapper>
           <SignUpTitle />
           <S.SignUpFormContainer>
             {signUpPage === 1 && <SignUpPage1 />}
             {signUpPage === 2 && <SignUpPage2 />}
             {signUpPage === 3 && <SignUpPage3 />}
           </S.SignUpFormContainer>
-        </S.SignUpWrapper>
+        </AuthWrapper>
       )}
     </>
   )
