+++ shared/common/src/pages/find/Pagination/ChangePasswordSuccess/style.ts
@@ -1,4 +1,4 @@
-import { LoginWrapper } from '../../../login/style'
+import { LoginWrapper } from '../../../auth/login/style'
 import styled from '@emotion/styled'
 
 export const SignUpSuccessWrapper = styled(LoginWrapper)``
