+++ shared/common/src/pages/index.ts
@@ -1,23 +1,22 @@
-export { default as LoginPage } from './login'
-export { default as SignUpPage } from './signUp'
-export { default as HomePage } from './home'
-export {
-  PostPage,
-  PostDetailPage,
-  PostWritePage,
-  NoticePage,
-  NoticeDetailPage,
-} from './post'
-export { default as MyPage } from './my'
-export { InquiryPage, InquiryDetailPage } from './inquiry'
-export {
-  LecturePage,
-  LectureDetailPage,
-  LectureApplyListPage,
-  LectureCreatePage,
-} from './lecture'
-export { default as ActivityListPage } from './activity/list'
 export { default as ActivityDetailPage } from './activity/detail'
+export { default as ActivityListPage } from './activity/list'
+export { LoginPage, SignUpPage } from './auth'
 export { default as ClubDetailPage } from './club/detail'
 export { default as StudentPage } from './club/student'
 export { default as FindPage } from './find'
+export { default as HomePage } from './home'
+export { InquiryDetailPage, InquiryPage } from './inquiry'
+export {
+  LectureApplyListPage,
+  LectureCreatePage,
+  LectureDetailPage,
+  LecturePage,
+} from './lecture'
+export { default as MyPage } from './my'
+export {
+  NoticeDetailPage,
+  NoticePage,
+  PostDetailPage,
+  PostPage,
+  PostWritePage,
+} from './post'
